Move waveform and corresponding test into its own package.
